475: [Sec] IDOR: /company/time-off-policies/<id>/users (#277)
* delete unused fields of TimeOffPolicyRelatedUserDto

* fix bug

* fix IDOR

* fix IDOR

+++ src/main/java/shamu/company/timeoff/repository/PaidHolidayUserRepository.java
@@ -1,5 +1,6 @@
 package shamu.company.timeoff.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 import org.springframework.data.jpa.repository.Query;
 import shamu.company.common.repository.BaseRepository;
@@ -31,7 +32,7 @@ public interface PaidHolidayUserRepository extends BaseRepository<PaidHolidayUse
           + "AND u.deleted_at IS NULL ",
       nativeQuery = true
   )
-  List<Long> findAllUserIdByCompanyId(Long companyId);
+  List<BigInteger> findAllUserIdByCompanyId(Long companyId);
 
   @Query(
       value = "SELECT * FROM paid_holidays_users WHERE company_id = ?1 AND user_id = ?2 "
